Break out stage0 attestation generation logic
This is done to make it available for testing the oak SDK, which requires attestation evidence to initialize
