+++ stage0/Cargo.toml
@@ -15,6 +15,7 @@ log = "*"
 linked_list_allocator = { version = "*", features = ["alloc_ref"] }
 oak_core = { path = "../oak_core", default-features = false }
 oak_dice = { workspace = true }
+oak_stage0_dice = { workspace = true }
 oak_linux_boot_params = { path = "../linux_boot_params" }
 oak_sev_guest = { workspace = true, features = ["rust-crypto"] }
 p256 = { version = "*", default-features = false, features = ["ecdsa"] }
